+++ aa_forum/models.py
@@ -134,9 +134,11 @@ class Board(models.Model):
         Category, related_name="boards", on_delete=models.CASCADE
     )
     name = models.CharField(max_length=254)
-
     slug = models.SlugField(max_length=254, unique=True, allow_unicode=True)
     description = models.TextField(blank=True)
+    discord_webhook = models.CharField(
+        blank=True, null=True, default=None, max_length=254
+    )
     parent_board = models.ForeignKey(
         "self",
         blank=True,
@@ -228,7 +230,6 @@ class Topic(models.Model):
 
     board = models.ForeignKey(Board, related_name="topics", on_delete=models.CASCADE)
     subject = models.CharField(max_length=254)
-
     slug = models.SlugField(max_length=254, unique=True, allow_unicode=True)
     is_sticky = models.BooleanField(default=False, db_index=True)
     is_locked = models.BooleanField(default=False, db_index=True)