+++ aa_forum/views/forum.py
@@ -4,7 +4,9 @@ Forum related views
 
 from typing import Optional
 
+import requests
 from app_utils.logging import LoggerAddTag
+from app_utils.urls import reverse_absolute
 
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
@@ -16,6 +18,7 @@ from django.db.models import Count, Exists, OuterRef, Prefetch, Q
 from django.http import HttpResponse, HttpResponseNotFound, HttpResponseRedirect
 from django.shortcuts import redirect, render
 from django.urls import reverse
+from django.utils.html import strip_tags
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 
@@ -338,16 +341,50 @@ def board_new_topic(
                 topic = Topic(board=board, subject=form.cleaned_data["subject"])
                 topic.save()
 
-                Message(
+                message = Message(
                     topic=topic,
                     user_created=request.user,
                     message=form.cleaned_data["message"],
-                ).save()
+                )
+                message.save()
 
             logger.info(
                 f'{request.user} started a new topic "{topic.subject}" '
                 f'in board "{board.name}"'
             )
+
+            # Send to webhook if one is configured
+            if board.discord_webhook is not None:
+                webhook_url = board.discord_webhook
+                topic_url = reverse_absolute(
+                    "aa_forum:forum_topic",
+                    args=[board.category.slug, board.slug, topic.slug],
+                )
+
+                embed = {
+                    "description": strip_tags(
+                        message.message[:250] + "..."
+                        if len(message.message) > 250
+                        else message.message
+                    ),
+                    "title": topic.subject,
+                    "url": topic_url,
+                }
+
+                data = {
+                    "content": _(
+                        f'**New Topic has been posted in board "{board.name}"**'
+                    ),
+                    # "username": "custom username",
+                    "embeds": [embed],
+                }
+
+                result = requests.post(webhook_url, json=data)
+
+                try:
+                    result.raise_for_status()
+                except requests.exceptions.HTTPError as err:
+                    logger.info(f"Discord Webhook Error: {err}")
 
             return redirect(
                 "aa_forum:forum_topic",